requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
beautifulsoup4
requests
python-dotenv
streamlit
